Fix external script

+++ entrypoint.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 if [ -f "/var/config-fpm/php.ini" ]; then
-    cp /var/config-fpm/nginx.conf /etc/php/7.0/fpm/php.ini
+    cp /var/config-fpm/php.ini /etc/php/7.0/fpm/php.ini
 fi
 
 if [ -d "/var/config-fpm/pool.d/" ]; then
@@ -15,13 +15,15 @@ fi
 if [ "${XDEBUG_ENABLED}" = "1" ]; then
     echo "zend_extension=xdebug.so" > /etc/php/7.0/fpm/conf.d/20-xdebug.ini
 else
-    rm /etc/php/7.0/fpm/conf.d/20-xdebug.ini
+    if [ -f "/etc/php/7.0/fpm/conf.d/20-xdebug.ini" ]; then
+        rm /etc/php/7.0/fpm/conf.d/20-xdebug.ini
+    fi
 fi
 
 if [ -f "/var/config-fpm/script.sh" ]; then
-    exec "/var/config-fpm/script.sh"
+    chmod 755 /var/config-fpm/script.sh
+    /var/config-fpm/script.sh
 fi
 
-if [ "$1" != '/usr/sbin/php-fpm7.0 --nodaemonize' ]; then
-    /bin/bash -l -c "$*"
-fi
+exec /usr/sbin/php-fpm7.0 --nodaemonize
+
